implementing findByTitle and findById methods

+++ frameworks/src/main/java/com/webapp/repository/AnnotationDataAccessImpl.java
@@ -4,9 +4,11 @@ import com.webapp.entity.Annotation;
 import com.webapp.repository.entity.AnnotationEntity;
 import com.webapp.repository.mapper.AnnotationMapper;
 import com.webapp.usecase.data_access.AnnotationDataAccess;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 
 @Component
@@ -17,6 +19,22 @@ public class AnnotationDataAccessImpl implements AnnotationDataAccess {
     public AnnotationDataAccessImpl(AnnotationRepository annotationRepository, AnnotationMapper annotationMapper) {
         this.annotationRepository = annotationRepository;
         this.annotationMapper = annotationMapper;
+    }
+
+    @Override
+    public List<Annotation> findByTitle(String title) {
+        List<AnnotationEntity> annotationEntityList = annotationRepository.findByTitle(title);
+        return annotationEntityList.stream()
+                .map(annotationMapper::toAnnotation)
+                .toList();
+    }
+
+    @Override
+    public Annotation findById(UUID uuid) {
+        Optional<AnnotationEntity> annotationEntity = annotationRepository.findById(uuid);
+        return annotationMapper.toAnnotation(annotationEntity
+                .orElseThrow(() -> new EntityNotFoundException("Annotation does not exist with given id"))
+        );
     }
 
     @Override
@@ -36,6 +54,7 @@ public class AnnotationDataAccessImpl implements AnnotationDataAccess {
 
     @Override
     public void delete(UUID id) {
+        if(!annotationRepository.existsById(id)) throw new RuntimeException("");
         annotationRepository.deleteById(id);
     }
 }
